Add tracing collector to fpx (#56)
* Switch from http to grpc

When trying to send traces to otelcol using http it kept returning 400 bad requests

* Add script to start a simple otel collector

* Commit transactions through the store

This allows us to do the error mapping in 1 place (even though it currently only maps to InternalError). This will also make it easier to switch to a generic TX implementation, if we're ever going to support multiple backends

* Work in progress implementation for the gRPC service

* Generate OTEL files during build

* Generate otel files up front and add tooling to generate it again

* Cleanup generated comments

* Intermediate

* Intermediate2

* Intermediate

* more:

* Replace generated code with public crate

* Working

* Support SpanKind

* Add support for start and end time

* Fix tests

* Add support for attributes

* Fix tests

* Add state

* Implement all trace fields

* Allow fetching of all spans for a single trace

* Resolve clippies

* Include serde for the otel types

* Refactor IntoResponse to ApiError

* Remove .gitmodules file

* Add schemars to OTEL models

* Naive implementation of http/json collector endpoint

Note this implementation comes with some major concurrency flaws

* Add decode as json or proto to otel http collector

* Naive lock to prevent any concurrency issues

* Move models module to api module

Resolve clippies

* Rejiggle some imports

* Rewrite conversion to more idiomatic Rust

Change storage of spans to include the API model

* Add instrumentation to Grpc's export fn

* Move grpc into the root module

Add service to only define the ingest logic once

* Update to opentelemetry

Cleanup dependencies (ie. remove unused features/crates)

* Implement readwrite/readonly transaction

This implementation still depends on a global lock, so it doesn't allow for concurrent reads. Trying to implement this lead to other issues. This is something that we should solve later.

* Remove TransactionBehavior::ReadOnly

Using this in start_readwrite_transaction was a accidental commit from a test to see if it works.

Remove both usages since it doesn't work for local sqlite connection and only seems to work for hosted libsql instances

If we have dedicated readonly connections, then we should be able to set the 'query_only' pragma to true, but since we only have 1 connection this won't work

* Sort features

This is a result of a reverted experiment, but I still wanted to include this change

* Add comments

* Add documentation and get rid of unwrap/expect

* Create a new span for each websocket connection

This allows us to add information to this as well, such as the ws_id. This also fixes the span propagation to the write loop

Update more docs

* Move some structs around, add some docs

* Fix tests and prevent compiler warnings

* Fix index

* Add documentation to data/util

* Simplify gRPC handler

* Run cargo update

* Use correct content type for the response (json already does this)

Update comment, related to spec

* Rewrite RUST_LOG parsing

* Allow grpc listen address to be changed

* Set NOT NULL rather than NULL