Partial attempt to support JsonSchema for Span
This was failing due to two issues:

No support for the time crate in schemars. There is a PR out, but it seems stuck. I also don't think this PR will support 'with = rfc3339' that we use

BTreeMap in Attribute map doesn't seem to be supported, did not dig into the reason for this issue