Refactor fpx to use the new fpx_lib (#112)
* Refactor fpx to use the new fpx_lib

Create implementation for Store and ServerEvents, both use their
original implementation with some caveats.

Delete all unused components. These components might be returned in
a future PR, but it is easier to now get rid of it, instead of having to
worry about backwards compatibility.

* Resolve clippies

* Partial attempt to support JsonSchema for Span

This was failing due to two issues:

No support for the time crate in schemars. There is a PR out, but it seems stuck. I also don't think this PR will support 'with = rfc3339' that we use

BTreeMap in Attribute map doesn't seem to be supported, did not dig into the reason for this issue

* Move api/mod.rs to api.rs

* Run 'xtask generate-schemas'