Sync Rust FPX workers code into the main branch (#130)
* intermediate

* First

* Make fpx-lib working

* Working

* Resolve clippies, compile issues

* Reimplement endpoints in fpx-workers (#103)

* Reimplement endpoints in fpx-workers

* Remove debug commands

* Replace OnceLock with LazyLock

Resolve clippies

* Resolve issue with cargo-deny

Make sure all crates use the same version

* Add rudimentary support for D1 (#107)

* Add rudimentary support for D1

* Resolve clippies

* Remove FakeStore

This Store impl was only used for proof of concept testing. Both fpx and fpx-worker should use their own working implementation

* Add support for SpanKind

* Tryout static dispatch

* Revert "Tryout static dispatch"

This reverts commit 9080a0cce74efc35b09e97cf9d07e29963abb8fe.

* Resolve issues with serialization

Ensure that Timestamp serializes as a fractional number representing the number of seconds since unix timestamp, as opposed to the number of nanoseconds since unix previously.

Serialize the content of Json<T> as a JSON encoded string as opposed to a object (when serializing a struct)

* FPX Worker Socket implementation (#105)

* Reimplement endpoints in fpx-workers

* Basics for hibernating websock server

* Fix routes on temporary router implementation

* Simple broadcasting

* Move ws to it's own module

* Forward broadcasting

* Move ws handlers to an axum router

* Reorganize

* Refactor

* Clean up pr

* Remove debug handler macro

* Use worker request for ws client

* Cleanup

* Add ServerEvents as a trait

* Fix imports

* Fix toml

* PR Feedback

* Doc comments

---------

Co-authored-by: Benno van den Berg <bennovandenberg@gmail.com>

* Resolve test issue and clippies (#110)

* Handle WebSocket close messages (#114)

* Refactor fpx to use the new fpx_lib (#112)

* Refactor fpx to use the new fpx_lib

Create implementation for Store and ServerEvents, both use their
original implementation with some caveats.

Delete all unused components. These components might be returned in
a future PR, but it is easier to now get rid of it, instead of having to
worry about backwards compatibility.

* Resolve clippies

* Partial attempt to support JsonSchema for Span

This was failing due to two issues:

No support for the time crate in schemars. There is a PR out, but it seems stuck. I also don't think this PR will support 'with = rfc3339' that we use

BTreeMap in Attribute map doesn't seem to be supported, did not dig into the reason for this issue

* Move api/mod.rs to api.rs

* Run 'xtask generate-schemas'

* Add production env for wrangler (#116)

* FPX Worker hibernating WebSocket restoration (#119)

* Restore hibernated websockets

* Environments are a bit picky, just use these for now

* Apply suggestions from code review

Remove commented code

Co-authored-by: Stephan Lagerwaard <stephanlagerwaard@gmail.com>

* Allow fetching the summary of a single trace (#113)

* Add support to get the summary for a single trace

* Add number of spans for a trace

Improve client tracing setup, Ok or Err is now set, including a error message based on the ClientError object

* Add support for listing traces

* Implement traces_list for D1Store

Resolve clippies

* Reorganize Cargo.toml

* Use 'use' for OtelLayer

* Support empty ApiError enums in the ApiError macro

Remove TraceListError::FakeError

* Actually sort by end_time

* Remove debug line

* Resolve clippy

* Remove unused variable

* Ensure that the number of spans is always > 0

Otherwise return None, which indicates that the trace was not able to be constructed

---------

Co-authored-by: Stephan Lagerwaard <stephanlagerwaard@gmail.com>